requirements.txt
fastapi[standard]
uvicorn[standard]
motor         # Driver asíncrono para MongoDB, recomendado para FastAPI
pydantic[email]
python-dotenv